services/auth-service/requirements.txt
# Authentication Service Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
